log in with pwd implemented

+++ src/main/java/com/Lord_vinayaka/web_mvc_module/repository/SignupRepo.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 import com.Lord_vinayaka.web_mvc_module.entity.UserProfile;
 @Repository
-public interface SignupRepo extends CrudRepository<UserProfile,Short> {
+public interface SignupRepo extends CrudRepository<UserProfile,Long> {
 
 }
